+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Improved iOS compat data - added mapping iOS 12.2 -> Safari 12.1, added bug fixes from patch releases
 - Added missed in `core-js-compat` helpers `ie_mob` normalization
 - Normalize the result of `getModulesListForTargetVersion` `core-js-compat` helper
+- Improved CI detection in the `postinstall` script, [#707](https://github.com/zloirock/core-js/issues/707)
 
 ##### 3.4.5 - 2019.11.28
 - Detect incorrect order of operations in `Object.assign`, MS Edge bug